fix hook 삭제
